+++ sdk/azcore/arm/runtime/pipeline.go
@@ -31,13 +31,23 @@ func NewPipeline(module, version string, cred azcore.TokenCredential, options *a
 		perCallPolicies = append(perCallPolicies, NewRPRegistrationPolicy(string(ep), cred, &regRPOpts))
 	}
 	perRetryPolicies := []policy.Policy{
-		azruntime.NewBearerTokenPolicy(cred, azruntime.AuthenticationOptions{
+		NewBearerTokenPolicy(cred, AuthenticationOptions{
 			TokenRequest: policy.TokenRequestOptions{
 				Scopes: []string{shared.EndpointToScope(string(ep))},
 			},
 			AuxiliaryTenants: options.AuxiliaryTenants,
-		},
-		),
+		}),
 	}
 	return azruntime.NewPipeline(module, version, perCallPolicies, perRetryPolicies, &options.ClientOptions)
+}
+
+// AuthenticationOptions contains various options used to create a credential policy.
+type AuthenticationOptions struct {
+	// TokenRequest is a TokenRequestOptions that includes a scopes field which contains
+	// the list of OAuth2 authentication scopes used when requesting a token.
+	// This field is ignored for other forms of authentication (e.g. shared key).
+	TokenRequest policy.TokenRequestOptions
+	// AuxiliaryTenants contains a list of additional tenant IDs to be used to authenticate
+	// in cross-tenant applications.
+	AuxiliaryTenants []string
 }